use float instead of int for activity value

+++ Moteur/src/main/kotlin/indicateur/Indicateur.kt
@@ -7,4 +7,4 @@ import java.util.*
 data class Result(var mean: MutableList<Activity>, var indicateurs: MutableMap<String, MutableList<Activity>>)
 
 @Serializable
-data class Activity(@Serializable(with = DateSerializer::class)val date: Date, var value: Int)
+data class Activity(@Serializable(with = DateSerializer::class)val date: Date, var value: Float)
